WebApp_MQTT_Client/requirements.txt
Flask
paho-mqtt
Flask-Login
Flask-SQLAlchemy
gunicorn
pytz
psycopg2-binary
